UberEat/backend/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
Faker==18.9.0
python-dotenv==1.0.0
